refactor(backend-cli): extract interfaces to src/types/database.ts
Follow the convention: types/interfaces live in `<pkg>/src/types/`, not
inline in action files. Also clarify this rule in AGENTS.md so future
agents place types correctly (package-local vs shared in
packages/common/src/types/).
