+++ AGENTS.md
@@ -87,7 +87,7 @@ For all TypeScript, JavaScript, and Vue files:
 
 - Write complete code blocks
 - ESM imports/exports for TypeScript
-- Interfaces in separate `src/types/` files | Utils in `src/utils/`
+- Types/interfaces: package-local in `<pkg>/src/types/` (app or package) | shared across backend+frontend in `packages/common/src/types/` — NEVER inline in action/component/util files. Utils in `src/utils/`.
 - Modularize into smaller reusable pieces
 - Document in code (not separate docs) | Reference existing docs, don't rewrite
 - English for comments/docs/code
