+++ apps/backend-cli/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './api-key.js';
+export * from './database.js';
 export * from './email.js';
 export * from './migrate-uploads.js';
 export * from './payment.js';
